# backend/docs/swagger.yaml
definitions:
...
  auth.LoginAPIResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      payload:
        $ref: '#/definitions/auth.LoginView'
    type: object
  auth.LoginView:
    properties:
      accessToken:
        type: string
      expiresAt:
        type: string
      isNewUser:
        type: boolean
      tokenType:
        type: string
      user:
        $ref: '#/definitions/auth.UserView'
    type: object
  auth.UserView:
    properties:
      creditBalance:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      referralCode:
        type: string
    type: object
...
  errors.APIError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: This is a sample server Petstore server.
  title: Gin APP API
  version: "1.0"
paths:
...
  /api/v1/auth/google/callback:
    get:
      parameters:
      - description: Google authorization code
        in: query
        name: code
        required: true
        type: string
      - description: OAuth state
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.APIError'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/errors.APIError'
      summary: Complete Google OAuth login
      tags:
      - Authentication
  /api/v1/auth/google/login:
    get:
      responses:
        "307":
          description: Temporary Redirect
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/errors.APIError'
      summary: Start Google OAuth login
      tags:
      - Authentication
...
securityDefinitions:
  APIAuth:
    in: cookie
    name: app_cookie_name
    type: apiKey
  OpenAPIAuth:
    in: header
    name: PRIVATE-TOKEN
    type: apiKey
swagger: "2.0"